Initial release
This is the first release of this fullstack-project. The repository does not have a proper name yet.

The index, script and styles files are linked. You can watch videos on the browser itself as the video is embedded on the webpage. There is a toggle for enabling dark mode as well. If a video is found interesting, you can add it to the (now persistent) watch later playlist. With future release, I will add the functionality to remove the videos, and make individual playlists too.